server/sm-pipeline-cdk/requirements.txt
aws-cdk-lib==2.66.1
constructs>=10.0.0,<11.0.0
aws_cdk.aws_lambda_python_alpha
shortuuid
requests
argparse